Reset and CLeanup modificcations for Cluster admin featue
